Fix some warnings

+++ source/nnpp.hpp
@@ -58,11 +58,11 @@ public:
 	inline auto begin() const { return m_array.begin(); }
 	inline auto end() const { return m_array.begin() + m_size; }
 	inline bool empty() const { return m_size == 0; }
-	inline void erase(uint index) { std::swap(m_array[index], m_array[--m_size]); }
+	inline void erase(ulong index) { std::swap(m_array[index], m_array[--m_size]); }
 	inline void clear() { m_size = 0; }
 	inline T& addNew() { return m_array[m_size++]; }
-	inline T& operator[](uint pos) { m_size = std::max(m_size, pos + 1); return m_array[pos]; }
-	inline const T& operator[](uint pos) const { return m_array[pos]; }
+	inline T& operator[](ulong pos) { m_size = std::max(m_size, pos + 1); return m_array[pos]; }
+	inline const T& operator[](ulong pos) const { return m_array[pos]; }
 
 	inline StackVector& operator=(StackVector&& other) {
 		m_array = std::move(other.m_array);
